Refactor: user 디렉토리 이름을 member로 변경
