+++ src/main/java/com/service/dida/domain/like/Like.java
@@ -1,7 +1,7 @@
 package com.service.dida.domain.like;
 
 import com.service.dida.domain.nft.Nft;
-import com.service.dida.domain.user.entity.Member;
+import com.service.dida.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
