+++ src/main/java/com/service/dida/domain/market/Market.java
@@ -2,7 +2,7 @@ package com.service.dida.domain.market;
 
 import com.service.dida.global.common.BaseEntity;
 import com.service.dida.domain.nft.Nft;
-import com.service.dida.domain.user.entity.Member;
+import com.service.dida.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
