+++ src/main/java/com/service/dida/domain/market/MarketService.java
@@ -6,8 +6,8 @@ import com.service.dida.global.config.exception.BaseException;
 import com.service.dida.domain.like.LikeRepository;
 import com.service.dida.domain.market.dto.*;
 import com.service.dida.domain.nft.Nft;
-import com.service.dida.domain.user.entity.Member;
-import com.service.dida.domain.user.repository.MemberRepository;
+import com.service.dida.domain.member.entity.Member;
+import com.service.dida.domain.member.repository.MemberRepository;
 import com.service.dida.global.util.UtilService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
