+++ src/main/java/com/service/dida/domain/member/dto/MemberRequestDto.java
@@ -1,4 +1,4 @@
-package com.service.dida.domain.user.dto;
+package com.service.dida.domain.member.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
