+++ src/main/java/com/service/dida/domain/member/service/GetMemberService.java
@@ -1,9 +1,9 @@
-package com.service.dida.domain.user.service;
+package com.service.dida.domain.member.service;
 
-import com.service.dida.domain.user.dto.SendAuthEmailDto;
-import com.service.dida.domain.user.entity.Member;
-import com.service.dida.domain.user.repository.MemberRepository;
-import com.service.dida.domain.user.usecase.GetMemberUseCase;
+import com.service.dida.domain.member.dto.SendAuthEmailDto;
+import com.service.dida.domain.member.entity.Member;
+import com.service.dida.domain.member.repository.MemberRepository;
+import com.service.dida.domain.member.usecase.GetMemberUseCase;
 import com.service.dida.global.config.exception.BaseException;
 import com.service.dida.global.config.exception.errorCode.UserErrorCode;
 import com.service.dida.global.util.mail.MailUseCase;
