+++ src/main/java/com/service/dida/domain/member/usecase/UpdateMemberUseCase.java
@@ -1,6 +1,6 @@
-package com.service.dida.domain.user.usecase;
+package com.service.dida.domain.member.usecase;
 
-import com.service.dida.domain.user.dto.MemberResponseDto;
+import com.service.dida.domain.member.dto.MemberResponseDto;
 import org.springframework.security.core.Authentication;
 
 public interface UpdateMemberUseCase {
