+++ src/main/java/com/service/dida/domain/post/service/PostService.java
@@ -7,8 +7,8 @@ import com.service.dida.domain.post.dto.EditPostReq;
 import com.service.dida.domain.post.dto.PostPostReq;
 import com.service.dida.domain.post.repository.PostRepository;
 import com.service.dida.domain.post.usecase.PostUseCase;
-import com.service.dida.domain.user.entity.Member;
-import com.service.dida.domain.user.repository.MemberRepository;
+import com.service.dida.domain.member.entity.Member;
+import com.service.dida.domain.member.repository.MemberRepository;
 import com.service.dida.global.config.exception.BaseException;
 import com.service.dida.global.config.exception.errorCode.MemberErrorCode;
 import com.service.dida.global.config.exception.errorCode.NftErrorCode;
